Observaciones con base de ACL
id_operador obtenido desde ACL
stencil default orderByDescId
Exceptions devuelven error_type

+++ app/Http/Controllers/ControlDeStencil/v1/Observaciones/ObservacionesAbm.php
@@ -6,10 +6,16 @@ use App\Http\Controllers\ControlDeStencil\v1\_model\Observaciones;
 use App\Http\Controllers\ControlDeStencil\v1\_request\ObservacionesCreateReq;
 use App\Http\Controllers\ControlDeStencil\v1\_request\ObservacionesUpdateReq;
 use App\Http\Controllers\Controller;
+use App\Http\Controllers\Core\ApiLogin;
 use Carbon\Carbon;
 
 class ObservacionesAbm extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth.api')->except('show');
+    }
+
     public function show($id) {
         $show = Observaciones::findOrFail($id);
         return $show;
@@ -19,6 +25,7 @@ class ObservacionesAbm extends Controller
     {
         $now = Carbon::now();
         $req->merge([
+	    'id_operador' => ApiLogin::user('id'),
             'fecha' => $now,
             'hora' => $now,
         ]);
@@ -44,7 +51,14 @@ class ObservacionesAbm extends Controller
     public function delete($id)
     {
         $item = Observaciones::findOrFail($id);
-        $deleted = $item->delete();
-        return compact('deleted');
+        if(
+            ApiLogin::owner($item->id_operador) ||
+            ApiLogin::isAdmin()
+        ) {
+            $deleted = $item->delete();
+            return compact('deleted');
+        }
+
+        return ['error'=>'No tiene permiso para eliminar este registro'];
     }
 }
